chore: isolate more util functions
